refactor(lowy): rename clipboard.ts to terminalScratch.ts
The module is named after its first consumer (browser clipboard
pastes) rather than the volatility axis it actually encapsulates:
per-terminal scratch storage for browser-originated content. After
the saveTerminalFile collapse, both that file and the surrounding
constants (koluClipboardDir, cleanupClipboardDir) are misleading —
the same machinery now also stores drag-and-drop file drops.

Renames:
- packages/server/src/clipboard.ts → terminalScratch.ts
- cleanupClipboardDir → cleanupTerminalScratch
- koluClipboardDir → koluScratchDir
- on-disk path component `clipboard/` → `scratch/` (under
  $XDG_RUNTIME_DIR/kolu-<pid>/; wiped on server shutdown, so the
  path-component rename is safe — no external consumers)

Browser-clipboard references in client code (ClipboardEvent,
ui/clipboard.ts, "Copied to clipboard" toasts) stay as-is — those
are legitimate clipboard semantics, not server scratch storage.
